requirements.txt
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.1
idna==3.3
pyshorteners==1.0.1
pyTelegramBotAPI==4.7.0
python-dotenv==0.20.0
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.12